src/code_synthesis/engine.py
"""Main synthesis engine orchestrating the code generation pipeline."""

from typing import Optional, Dict, Any
import yaml
import os


class SynthesisEngine:
    """
    Main engine for generating code from natural language descriptions.
    
    Pipeline:
    1. Parse natural language into structured intent
    2. Retrieve relevant patterns from codebase knowledge base
    3. Generate code using local LLM with context
    4. Validate and format output
    """
    
    def __init__(self, config_path: Optional[str] = None):
        """
        Initialize the synthesis engine.
        
        Args:
            config_path: Path to config.yaml. If None, uses default.
        """
        self.config = self._load_config(config_path)
        self._parser = None
        self._pattern_matcher = None
        self._generator = None
        self._validator = None
        
    def _load_config(self, config_path: Optional[str]) -> Dict[str, Any]:
        """Load configuration from YAML file."""
        if config_path is None:
            # Look for config in standard locations
            candidates = [
                "config.yaml",
                os.path.expanduser("~/.config/code-synthesis/config.yaml"),
                "/etc/code-synthesis/config.yaml"
            ]
            for candidate in candidates:
                if os.path.exists(candidate):
                    config_path = candidate
                    break
        
        if config_path and os.path.exists(config_path):
            with open(config_path, 'r') as f:
                return yaml.safe_load(f)
        
        # Return default config
        return {
            "model": {
                "temperature": 0.2,
                "max_tokens": 2048
            },
            "patterns": {
                "similarity_threshold": 0.85,
                "max_patterns": 5
            }
        }
    
    def generate(self, description: str, language: str = "python") -> str:
        """
        Generate code from a natural language description.
        
        Args:
            description: Natural language description of desired code
            language: Target programming language (default: python)
            
        Returns:
            Generated and validated code as string
            
        Example:
            >>> engine = SynthesisEngine()
            >>> code = engine.generate(
            ...     "Create a function to calculate fibonacci numbers recursively"
            ... )
        """
        # TODO: Implement full pipeline
        # 1. Parse intent
        # 2. Match patterns
        # 3. Generate with LLM
        # 4. Validate
        
        return self._placeholder_generate(description, language)
    
    def _placeholder_generate(self, description: str, language: str) -> str:
        """Placeholder implementation - returns template response."""
        return f"""# Generated code for: {description[:50]}...
# Language: {language}
# Status: Foundation laid - full implementation pending

# TODO: Integrate with local LLM for actual generation
# TODO: Add pattern matching from codebase
# TODO: Add validation pipeline

def placeholder_function():
    \"\"\"
    This is a placeholder. The full synthesis engine will:
    1. Parse your description into structured requirements
    2. Find similar patterns in your codebase
    3. Generate code using local LLM
    4. Validate syntax and types
    \"\"\"
    pass
"""
    
    def index_codebase(self, path: str) -> None:
        """
        Index a codebase to learn patterns from it.
        
        Args:
            path: Root path of the codebase to index
        """
        # TODO: Implement indexing
        raise NotImplementedError("Codebase indexing not yet implemented")
    
    def get_status(self) -> Dict[str, Any]:
        """Get current engine status and configuration."""
        return {
            "version": "0.1.0",
            "status": "foundation_laid",
            "components": {
                "parser": "not_initialized",
                "pattern_matcher": "not_initialized", 
                "generator": "not_initialized",
                "validator": "not_initialized"
            },
            "config": self.config
        }
